+++ src/@components/@common/hooks/useRecentlyBookmarked.ts
@@ -1,9 +1,9 @@
 import useSWR from "swr";
 
-import { realReq } from "../../../../core/api/common/axios";
-import { PATH } from "../../../../core/api/common/constants";
-import { RecentCardList } from "../../../../types/cardCollection";
-import { PiickleSWRResponse } from "../../../../types/remote/swr";
+import { realReq } from "../../../core/api/common/axios";
+import { PATH } from "../../../core/api/common/constants";
+import { RecentCardList } from "../../../types/cardCollection";
+import { PiickleSWRResponse } from "../../../types/remote/swr";
 
 export function useRecentlyBookmarked() {
   const { data } = useSWR<PiickleSWRResponse<RecentCardList>>(`${PATH.CARDS_}${PATH.CARDS_RECENT}`, realReq.GET_SWR, {
