+++ src/@components/BestPiicklePage/BestPiickleRecommend/index.tsx
@@ -2,7 +2,7 @@ import { CardList, LocationType } from "../../../types/cardCollection";
 import { HeadingTitle } from "../../../util/main/headingTitles";
 import HeadingTitleContainer from "../../@common/HeadingTitleContainer";
 import { useCardsByGender } from "./hooks/useCardsByGender";
-import { useRecentlyBookmarked } from "./hooks/useRecentlyBookmarked";
+import { useRecentlyBookmarked } from "../../@common/hooks/useRecentlyBookmarked";
 import RecommendItem from "./RecommendItem";
 import * as St from "./style";
